requirements.txt
torch>=1.9.0
numpy>=1.21.0
scipy>=1.7.0
openai-whisper>=20231117 